Fixed sync_and_restart_server script so that it works when run locally on the VM.
PRESUBMIT=passed
R=asharif,bjanakiraman
DELTA=6  (0 added, 1 deleted, 5 changed)
OCL=46038-p2
RCL=46046-p2
RDATE=2010/12/21 08:48:06


P4 change: 42607492

+++ v14/automation/server/sync_and_restart_server.sh
@@ -1,13 +1,12 @@
-become mobiletc-prebuild@chrome-dev1.hot <<EOF
+#!/bin/bash
 PID=$(ps ax | grep server.py | grep -v grep | cut -d' ' -f 2)
-###if ! [[ -z $PID ]]; then
+if ! [[ -z $PID ]]; then
   kill $PID
   sleep 5
   kill -9 $PID
-###fi
+fi
 cd ~/perforce2/gcctools/chromeos/v14
 g4 sync ...
 cd automation/server
-pwd
-nohup python server.py -m test_pool.csv &
-EOF
+export PYTHONPATH=../..:$PYTHONPATH
+(nohup python2.6 server.py -m test_pool.csv &) </dev/null 1>/dev/null 2>/dev/null </dev/null
